move exception and tests
