+++ mt-data/src/test/java/org/zanata/magpie/model/LocaleTest.java
