+++ mt-data/src/test/java/org/zanata/magpie/model/type/LocaleTypeTest.java
