+++ server/src/test/java/org/zanata/magpie/service/TranslatableHTMLNodeTest.java
@@ -1,4 +1,4 @@
-package org.zanata.magpie.model;
+package org.zanata.magpie.service;
 
 import org.assertj.core.util.Lists;
 import org.jsoup.nodes.Attributes;
@@ -6,7 +6,6 @@ import org.jsoup.nodes.Element;
 import org.jsoup.nodes.Node;
 import org.jsoup.parser.Tag;
 import org.junit.Test;
-import org.zanata.magpie.service.TranslatableHTMLNode;
 
 import java.util.HashMap;
 import java.util.Map;
